+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/models/DosageFormula.java
@@ -1,4 +1,4 @@
-package fi.tuni.koodimankelit.antibiootit.calculator.data;
+package fi.tuni.koodimankelit.antibiootit.models;
 
 public class DosageFormula {
