+++ backend/antibiootit/src/main/java/fi/tuni/koodimankelit/antibiootit/models/Measurement.java
@@ -1,4 +1,4 @@
-package fi.tuni.koodimankelit.antibiootit.calculator.data;
+package fi.tuni.koodimankelit.antibiootit.models;
 
 
 public class Measurement {
